requirements.txt
anyio==4.0.0
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==1.2.1
exceptiongroup==1.1.3
Faker==19.3.1
fastapi==0.103.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.0
pydantic==1.10.12
pymongo==3.12.0
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.15.0
